Update index.html - adding window.onload function
This will ensure the data (especially on the dashboard) will load from local storage. Prior to this change, if I enter some data (steps or weight) it logs it correctly on the Tracking or Progress tab and even displays it correctly on the dashboard page. But if I do  refresh, it disappears from the dashboard page